+++ src/library/xlib/xevents.h
@@ -46,6 +46,8 @@ OVERRIDE int XIfEvent(Display *display, XEvent *event_return, Bool (*predicate)(
 OVERRIDE Bool XCheckIfEvent(Display *display, XEvent *event_return, Bool (*predicate)(Display *, XEvent *, XPointer), XPointer arg);
 OVERRIDE Status XSendEvent(Display *display, Window w, Bool propagate, long event_mask, XEvent *event_send);
 
+OVERRIDE Bool XFilterEvent(XEvent *event, Window w);
+
 OVERRIDE int XFlush(Display *display);
 OVERRIDE int XSync(Display *display, Bool discard);