+++ CHANGELOG.md
@@ -48,6 +48,7 @@
 * Change audio context locking
 * Fix bug when accessing samples of empty buffer
 * Fix bug when accessing past the audio buffer (#463)
+* Disable xevent filtering used for composition (#443)
 
 ## [1.4.2] - 2021-07-06
 ### Added